Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/xhr/send-blob-with-no-mime-type-expected.txt


FAIL Synchronous blob loading with no mime type [POST] assert_equals: expected "NO" but got ""
FAIL Asynchronous blob loading with no mime type [POST] assert_equals: expected "NO" but got ""
FAIL Synchronous blob loading with no mime type [PUT] assert_equals: expected "NO" but got ""
FAIL Asynchronous blob loading with no mime type [PUT] assert_equals: expected "NO" but got ""
FAIL Synchronous blob loading with invalid mime type [POST] assert_equals: expected "NO" but got ""
FAIL Asynchronous blob loading with invalid mime type [POST] assert_equals: expected "NO" but got ""
FAIL Synchronous blob loading with invalid mime type [PUT] assert_equals: expected "NO" but got ""
FAIL Asynchronous blob loading with invalid mime type [PUT] assert_equals: expected "NO" but got ""
